refactor(docs): split playground codec into contract, validation, and wire
Separate share-size/envelope validation from the #play=v1 base64url wire
format so each concern can change without sharing a 360-line module.
Public imports stay on playground-codec; characterization and facade
surface tests lock the stable API.
