+++ scripts/playground-codec.test.ts
@@ -8,7 +8,11 @@ import {
   encodePlaygroundSeed,
   PLAYGROUND_MAX_DECODED_BYTES,
   PLAYGROUND_MAX_ENCODED_LENGTH,
+  PLAYGROUND_MAX_FIELDS,
+  PLAYGROUND_MAX_ROWS,
+  PLAYGROUND_MAX_STRING_CODE_POINTS,
   PlaygroundCodecError,
+  validatePlaygroundSeed,
   type PlaygroundSeedV1,
 } from "../apps/docs/src/lib/playground-codec";
 
@@ -190,6 +194,135 @@ describe("playground fragment codec", () => {
       expect(result.error).toBeInstanceOf(PlaygroundCodecError);
       expect(result.error.code).toBe("INVALID_SPEC");
       expect(result.error.diagnostics[0]?.code).toBe("missing-required-channel");
+    }
+  });
+
+  test("rejects invalid JSON payloads after base64 decode", () => {
+    const binary = "{not-json";
+    const payload = btoa(binary).replaceAll("+", "-").replaceAll("/", "_").replace(/=+$/u, "");
+    expectCode(`#play=v1.${payload}`, "INVALID_JSON");
+  });
+
+  test("rejects unknown envelope fields on decode and validate", () => {
+    const withExtra = { ...seed(), unexpected: true };
+    expectCode(rawHash(withExtra), "INVALID_SEED");
+    expect(() => validatePlaygroundSeed(withExtra as PlaygroundSeedV1)).toThrow(
+      new PlaygroundCodecError(
+        "INVALID_SEED",
+        "Shared playground state contains unknown envelope fields.",
+      ),
+    );
+  });
+
+  test("accepts seeds at exact share limits and fails one over", () => {
+    // Column form stays under the JSON byte ceiling at the row limit.
+    const atRowsSeed: PlaygroundSeedV1 = {
+      version: 1,
+      source: { kind: "sample", id: "starter" },
+      spec: {
+        edition: 1,
+        data: {
+          columns: {
+            city: Array.from({ length: PLAYGROUND_MAX_ROWS }, () => "a"),
+            value: Array.from({ length: PLAYGROUND_MAX_ROWS }, (_, value) => value),
+          },
+        },
+        layers: spec.layers,
+      },
+    };
+    expect(validatePlaygroundSeed(atRowsSeed).spec.data).toEqual(atRowsSeed.spec.data);
+    expect(() =>
+      validatePlaygroundSeed({
+        ...atRowsSeed,
+        spec: {
+          ...atRowsSeed.spec,
+          data: {
+            columns: {
+              city: Array.from({ length: PLAYGROUND_MAX_ROWS + 1 }, () => "a"),
+              value: Array.from({ length: PLAYGROUND_MAX_ROWS + 1 }, (_, value) => value),
+            },
+          },
+        },
+      }),
+    ).toThrow(
+      new PlaygroundCodecError(
+        "TOO_MANY_ROWS",
+        `Shared playground state may contain at most ${String(PLAYGROUND_MAX_ROWS)} inline rows in total.`,
+      ),
+    );
+
+    const fieldNames = Array.from(
+      { length: PLAYGROUND_MAX_FIELDS },
+      (_, index) => `f${String(index)}`,
+    );
+    const atFieldsSeed: PlaygroundSeedV1 = {
+      version: 1,
+      source: { kind: "custom" },
+      spec: {
+        edition: 1,
+        data: {
+          columns: Object.fromEntries(fieldNames.map((name, index) => [name, [index]])),
+        },
+        layers: [
+          {
+            geom: "point",
+            stat: "identity",
+            position: "identity",
+            aes: { x: { field: "f0" }, y: { field: "f1" } },
+          },
+        ],
+      },
+    };
+    expect(validatePlaygroundSeed(atFieldsSeed).spec.data).toEqual(atFieldsSeed.spec.data);
+
+    const titleAtLimit = "a".repeat(PLAYGROUND_MAX_STRING_CODE_POINTS);
+    expect(
+      validatePlaygroundSeed(seed({ ...spec, labs: { title: titleAtLimit } })).spec.labs?.title,
+    ).toBe(titleAtLimit);
+  });
+
+  test("validatePlaygroundSeed prefers structural reasons over share-size ceilings", () => {
+    // Oversized string is reported before any whole-envelope DECoded/ENCODED ceiling.
+    const oversizedTitle = "🧭".repeat(PLAYGROUND_MAX_STRING_CODE_POINTS + 1);
+    expect(() =>
+      validatePlaygroundSeed(seed({ ...spec, labs: { title: oversizedTitle } })),
+    ).toThrow(
+      new PlaygroundCodecError(
+        "STRING_TOO_LONG",
+        `Shared playground strings must contain at most ${String(PLAYGROUND_MAX_STRING_CODE_POINTS)} Unicode code points.`,
+      ),
+    );
+
+    const tooManyRows = seed({
+      ...spec,
+      data: {
+        values: Array.from({ length: PLAYGROUND_MAX_ROWS + 1 }, (_, value) => ({ value })),
+      },
+    });
+    try {
+      validatePlaygroundSeed(tooManyRows);
+      throw new Error("expected validatePlaygroundSeed to throw");
+    } catch (error) {
+      expect(error).toBeInstanceOf(PlaygroundCodecError);
+      if (error instanceof PlaygroundCodecError) {
+        expect(error.code).toBe("TOO_MANY_ROWS");
+      }
+    }
+  });
+
+  test("validatePlaygroundSeed returns a bounded seed and preserves instanceof errors", () => {
+    const bounded = validatePlaygroundSeed(seed());
+    expect(bounded).toEqual(seed());
+    expect(bounded).not.toBe(seed());
+    try {
+      validatePlaygroundSeed(seed({ ...spec, data: { values: [{ value: 1 }] } }));
+      throw new Error("expected invalid spec");
+    } catch (error) {
+      expect(error).toBeInstanceOf(PlaygroundCodecError);
+      if (error instanceof PlaygroundCodecError) {
+        expect(error.name).toBe("PlaygroundCodecError");
+        expect(error.code).toBe("INVALID_SPEC");
+      }
     }
   });
 });
